[KAIZEN-0] flytte ut wireMockGateway til eget object som kan brukes av flere
For å sette opp flere tester med wireMock så flyttes denne logikken ut av SakApiGatewayTest
