+++ web/src/test/java/no/nav/modiapersonoversikt/utils/WireMockUtils.kt
@@ -1,0 +1,41 @@
+package no.nav.modiapersonoversikt.utils
+
+import com.github.tomakehurst.wiremock.WireMockServer
+import com.github.tomakehurst.wiremock.client.MappingBuilder
+import com.github.tomakehurst.wiremock.client.WireMock
+
+object WireMockUtils {
+    fun withMockGateway(
+        stub: WireMockServer.() -> Unit = { },
+        verify: ((WireMockServer) -> Unit),
+        test: (String) -> Unit
+    ) {
+        val wireMockServer = WireMockServer()
+        try {
+            stub(wireMockServer)
+            wireMockServer.start()
+            test("http://localhost:${wireMockServer.port()}")
+            verify(wireMockServer)
+        } finally {
+            wireMockServer.stop()
+        }
+    }
+
+    fun getWithBody(statusCode: Int = 200, body: String? = null): WireMockServer.() -> Unit = {
+        this.stubFor(WireMock.get(WireMock.anyUrl()).withBody(statusCode, body))
+    }
+
+    fun postWithBody(statusCode: Int = 200, body: String? = null): WireMockServer.() -> Unit = {
+        this.stubFor(WireMock.post(WireMock.anyUrl()).withBody(statusCode, body))
+    }
+
+    private fun MappingBuilder.withBody(statusCode: Int = 200, body: String? = null): MappingBuilder {
+        this.willReturn(
+            WireMock.aResponse()
+                .withStatus(statusCode)
+                .withHeader("Content-Type", "application/json")
+                .withBody(body)
+        )
+        return this
+    }
+}
